Refactor package layout
Moved coh event action classes out of controller package into their own
package.
