+++ src/main/java/uk/gov/hmcts/reform/sscscorbackend/coheventmapper/actions/DeadlineExtendedEventAction.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscscorbackend.controllers.coheventmapper;
+package uk.gov.hmcts.reform.sscscorbackend.coheventmapper.actions;
 
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
