+++ src/main/java/uk/gov/hmcts/reform/sscscorbackend/coheventmapper/actions/DeadlineExtensionDeniedEventAction.java
@@ -1,14 +1,14 @@
-package uk.gov.hmcts.reform.sscscorbackend.controllers.coheventmapper;
+package uk.gov.hmcts.reform.sscscorbackend.coheventmapper.actions;
 
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
 import uk.gov.hmcts.reform.sscscorbackend.service.pdf.CohEventActionContext;
 
 @Service
-public class DeadlineExtensionGrantedEventAction implements CohEventAction {
+public class DeadlineExtensionDeniedEventAction implements CohEventAction {
     @Override
     public String cohEvent() {
-        return "question_deadline_extension_granted";
+        return "question_deadline_extension_denied";
     }
 
     @Override
@@ -18,12 +18,11 @@ public class DeadlineExtensionGrantedEventAction implements CohEventAction {
 
     @Override
     public EventType getCcdEventType() {
-        return EventType.COH_QUESTION_DEADLINE_EXTENSION_GRANTED;
+        return EventType.COH_QUESTION_DEADLINE_EXTENSION_DENIED;
     }
 
     @Override
     public boolean notifyAppellant() {
         return false;
     }
-
 }
