+++ src/main/java/uk/gov/hmcts/reform/sscscorbackend/coheventmapper/actions/DecisionRejectedEventAction.java
@@ -1,14 +1,14 @@
-package uk.gov.hmcts.reform.sscscorbackend.controllers.coheventmapper;
+package uk.gov.hmcts.reform.sscscorbackend.coheventmapper.actions;
 
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
 import uk.gov.hmcts.reform.sscscorbackend.service.pdf.CohEventActionContext;
 
 @Service
-public class DeadlineExtensionDeniedEventAction implements CohEventAction {
+public class DecisionRejectedEventAction implements CohEventAction {
     @Override
     public String cohEvent() {
-        return "question_deadline_extension_denied";
+        return "decision_rejected";
     }
 
     @Override
@@ -18,7 +18,7 @@ public class DeadlineExtensionDeniedEventAction implements CohEventAction {
 
     @Override
     public EventType getCcdEventType() {
-        return EventType.COH_QUESTION_DEADLINE_EXTENSION_DENIED;
+        return EventType.COH_DECISION_REJECTED;
     }
 
     @Override
