+++ src/main/java/uk/gov/hmcts/reform/sscscorbackend/controllers/OnlineHearingController.java
@@ -7,7 +7,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import uk.gov.hmcts.reform.sscscorbackend.controllers.coheventmapper.CohEventActionMapper;
+import uk.gov.hmcts.reform.sscscorbackend.coheventmapper.CohEventActionMapper;
 import uk.gov.hmcts.reform.sscscorbackend.service.OnlineHearingService;
 import uk.gov.hmcts.reform.sscscorbackend.thirdparty.ccd.apinotifications.CcdEvent;
 import uk.gov.hmcts.reform.sscscorbackend.thirdparty.coh.apinotifications.CohEvent;
