+++ src/test/java/uk/gov/hmcts/reform/sscscorbackend/coheventmapper/action/DeadlineElapsedEventActionTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscscorbackend.controllers.coheventmapper;
+package uk.gov.hmcts.reform.sscscorbackend.coheventmapper.action;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -7,20 +7,21 @@ import static org.mockito.Mockito.*;
 import org.junit.Test;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseDetails;
+import uk.gov.hmcts.reform.sscscorbackend.coheventmapper.actions.DeadlineElapsedEventAction;
 import uk.gov.hmcts.reform.sscscorbackend.service.CorEmailService;
 import uk.gov.hmcts.reform.sscscorbackend.service.EmailMessageBuilder;
-import uk.gov.hmcts.reform.sscscorbackend.service.StoreAnswersPdfService;
+import uk.gov.hmcts.reform.sscscorbackend.service.StoreAnswersDeadlineElapsedPdfService;
 import uk.gov.hmcts.reform.sscscorbackend.service.pdf.CohEventActionContext;
 import uk.gov.hmcts.reform.sscscorbackend.service.pdf.Pdf;
 
-public class AnswerSubmittedEventActionTest {
+public class DeadlineElapsedEventActionTest {
     @Test
     public void canSendPdf() {
         CorEmailService corEmailService = mock(CorEmailService.class);
-        StoreAnswersPdfService storeAnswersPdfService = mock(StoreAnswersPdfService.class);
+        StoreAnswersDeadlineElapsedPdfService storeAnswersPdfService = mock(StoreAnswersDeadlineElapsedPdfService.class);
         EmailMessageBuilder emailMessageBuilder = mock(EmailMessageBuilder.class);
 
-        AnswerSubmittedEventAction answerSubmittedEventAction = new AnswerSubmittedEventAction(corEmailService, storeAnswersPdfService, emailMessageBuilder);
+        DeadlineElapsedEventAction deadlineElapsedEventAction = new DeadlineElapsedEventAction(corEmailService, storeAnswersPdfService, emailMessageBuilder);
 
         String someCaseReference = "someCaseReference";
         SscsCaseDetails caseDetails = SscsCaseDetails.builder()
@@ -32,7 +33,7 @@ public class AnswerSubmittedEventActionTest {
         CohEventActionContext cohEventActionContext = new CohEventActionContext(new Pdf(new byte[]{2, 5, 6, 0, 1}, pdfName), caseDetails);
         when(emailMessageBuilder.getAnswerMessage(caseDetails)).thenReturn("some message");
 
-        CohEventActionContext result = answerSubmittedEventAction.handle(caseId, onlineHearingId, cohEventActionContext);
+        CohEventActionContext result = deadlineElapsedEventAction.handle(caseId, onlineHearingId, cohEventActionContext);
 
         verify(corEmailService).sendPdfToDwp(cohEventActionContext, "Appellant has provided information (" + someCaseReference + ")", "some message");
         assertThat(result, is(cohEventActionContext));
