+++ src/test/java/uk/gov/hmcts/reform/sscscorbackend/controllers/OnlineHearingControllerTest.java
@@ -11,7 +11,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import uk.gov.hmcts.reform.sscscorbackend.controllers.coheventmapper.CohEventActionMapper;
+import uk.gov.hmcts.reform.sscscorbackend.coheventmapper.CohEventActionMapper;
 import uk.gov.hmcts.reform.sscscorbackend.service.OnlineHearingService;
 import uk.gov.hmcts.reform.sscscorbackend.thirdparty.ccd.apinotifications.CcdEvent;
 import uk.gov.hmcts.reform.sscscorbackend.thirdparty.coh.apinotifications.CohEvent;
